requirements.txt
# Django Core & Production Optimized for Render 2025
# Python 3.13.4 Compatible Versions
Django==5.1.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-decouple==3.8

# Static Files & WhiteNoise for Render
whitenoise==6.7.0
django-storages==1.14.4

# Database - Python 3.13+ COMPATIBLE
psycopg[binary]==3.2.3
dj-database-url==2.2.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
cryptography==42.0.8

# AWS Services (Face Recognition, IoT)
boto3==1.35.76
botocore==1.35.76

# Face Recognition (Local Fallback - Optional)
# deepface==0.0.93
# opencv-python==4.10.0.84
# tf-keras==2.16.0

# API Documentation
drf-spectacular==0.27.2

# Payment Processing - Compatible version
stripe==10.11.0

# File Processing
Pillow==10.4.0

# HTTP Clients
requests==2.32.3
httpx==0.25.2

# Date/Time Utilities
python-dateutil==2.8.2

# Email Service
sendgrid==6.11.0

# PDF Generation
reportlab==4.0.7

# Monitoring
sentry-sdk==1.39.1
flower==2.0.1

# Environment
python-dotenv==1.0.1

# Production Server
gunicorn==22.0.0

# Task Queue & Async Processing
celery[redis]==5.3.4
django-celery-beat==2.5.0
redis==5.0.1

# WebSocket Support (Django Channels)
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# QuickBooks Integration - DISABLED FOR INITIAL DEPLOY  
# Will be enabled after successful basic deploy
# intuitlib==1.2.4
# requests-oauthlib==2.0.0
